+++ frontend/src/Router/Main.js
@@ -10,7 +10,7 @@ function Main(){
       <Switch>
         { /*<Route exact path='/' component={Home}></Route> */ }
         <Route exact path='/' render={(props) => <Landing {...props} />}></Route>
-        <Route exact path='/feed' render={(props) => <MainFeed {...props} />}/*component={MainFeed}*/></Route>
+        <Route exact path='/feed' render={(props) => <MainFeed {...props} />}></Route>
         <Route exact path='/attention' render={(props) => <MainAttentionPage {...props} />}></Route>
         <Route exact path='/error' render={(props) => <ErrorLanding {...props} />}></Route>
       </Switch>
@@ -18,4 +18,3 @@ function Main(){
   }
   
   export default Main;
-
