+++ frontend/src/components/Carousel/CarouselComponent/BtnSlider.js
