+++ frontend/src/components/Carousel/CarouselComponent/Carousel.css
