+++ frontend/src/components/Carousel/CarouselComponent/CarouselComponent.js
