+++ frontend/src/components/Carousel/CarouselComponent/SliderData.js
