--- frontend/src/components/Carousel/CarouselModal.css
@@ -1,17 +1,0 @@
-/*.modal {
-    position: fixed;
-    height: 450px;
-    width: 500px;
-    top: 50%;
-    left: 50%;
-    width: 50%;
-    min-width: 50%;
-    height: 60%;
-    min-height: 65%;
-    transform: translate(-50%, -50%);
-    background-color: grey;
-    /*border-radius: 2%;*/
-    /*padding: 2%;
-    border-style: solid;
-    border-color: black;
-  }*/
